refactor(ui): add RowAppendableTableModel interface
