+++ src/main/java/dev/donutquine/editor/gui/layout/components/tables/RowAppendableTableModel.java
@@ -1,0 +1,5 @@
+package dev.donutquine.editor.gui.layout.components.tables;
+
+public interface RowAppendableTableModel {
+    boolean isAppendRow(int row);
+}
